+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [a0.3.3] - unreleased
+### Added
+- Estimate the next N seconds, beside estimate the rest
+
 ## [a0.3.2] - 2026-09-25
 ### Added
 - Estimate the rest by the sound: boundaries placed in the pauses of the waveform, in books and videos